Created LaserGui and moved all fx handle here.
